requirements.txt
selenium==4.8.3
pandas==1.3.5
beautifulsoup4==4.12.3
html5lib==1.1
webdriver-manager==3.8.6
# Postgres driver
psycopg2-binary==2.9.7
# lxml is optional because html5lib is used by pandas.read_html
# lxml==4.9.3

# pip install -r requirements.txt
# python main.py        # 크롤링 실행
# python homerun.py     # 홈런 분석 (크롤링 후)
python-dotenv==1.0.0
